// web/src/api/models/system-default-quotas.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * System default quota configuration
 * @export
 * @interface SystemDefaultQuotas
 */
export interface SystemDefaultQuotas {
    /**
     * Default maximum collection count
     * @type {number}
     * @memberof SystemDefaultQuotas
     */
    'max_collection_count': number;
    /**
     * Default maximum document count
     * @type {number}
     * @memberof SystemDefaultQuotas
     */
    'max_document_count': number;
    /**
     * Default maximum documents per collection
     * @type {number}
     * @memberof SystemDefaultQuotas
     */
    'max_document_count_per_collection': number;
    /**
     * Default maximum bot count
     * @type {number}
     * @memberof SystemDefaultQuotas
     */
    'max_bot_count': number;
}